@@ -1,15 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
+    compileSdkVersion 31
     defaultConfig {
         applicationId "com.laioffer.tinnews"
         minSdkVersion 17
-        targetSdkVersion 26
+        targetSdkVersion 31
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
         vectorDrawables.useSupportLibrary = true
+        multiDexEnabled true
     }
 
     buildTypes {
@@ -20,38 +21,44 @@ android {
     }
 
     compileOptions {
-        sourceCompatibility JavaVersion.VERSION_1_8
-        targetCompatibility JavaVersion.VERSION_1_8
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
     }
+    namespace 'com.laioffer.tinnews'
 }
 
 ext {
-    googleVersion = '26.1.0'
+    googleVersion = '27.1.0'
     retrofitVersion = '2.2.0'
     rxjavaVersion = '2.0.1'
     room_version = '1.1.1'
 }
 
 dependencies {
+    implementation 'androidx.constraintlayout:constraintlayout:2.1.3'
+    implementation 'androidx.multidex:multidex:2.0.1'
+    debugImplementation 'androidx.fragment:fragment-testing:1.4.0-alpha03'
+    implementation 'com.google.android.material:material:1.2.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation "com.android.support:appcompat-v7:${googleVersion}"
+    implementation 'androidx.appcompat:appcompat:1.0.0'
     implementation 'com.facebook.stetho:stetho:1.5.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    implementation "com.android.support:support-v4:${googleVersion}"
-    implementation "com.android.support:design:${googleVersion}"
-    implementation "com.android.support:recyclerview-v7:${googleVersion}"
-    implementation "com.android.support:support-v4:${googleVersion}"
-    implementation 'com.android.support:support-v4:26.1.0'
-    implementation "android.arch.persistence.room:runtime:$room_version"
-    annotationProcessor "android.arch.persistence.room:compiler:$room_version"
-    implementation "android.arch.persistence.room:rxjava2:$room_version"
-    implementation "android.arch.persistence.room:guava:$room_version"
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.room:room-runtime:2.0.0'
+    annotationProcessor 'androidx.room:room-compiler:2.0.0'
+    implementation 'androidx.room:room-rxjava2:2.0.0'
+    implementation 'androidx.room:room-guava:2.0.0'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.1.0'
     implementation 'com.mindorks:placeholderview:0.7.1'
     implementation 'com.squareup.picasso:picasso:2.71828'
-    implementation "com.android.support:cardview-v7:${googleVersion}"
+    implementation 'androidx.cardview:cardview:1.0.0'
     implementation "io.reactivex.rxjava2:rxjava:${rxjavaVersion}"
     implementation "io.reactivex.rxjava2:rxandroid:${rxjavaVersion}"
     implementation "com.squareup.retrofit2:retrofit:${retrofitVersion}"
