@@ -1,11 +1,10 @@
 package com.laioffer.tinnews.common;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import com.laioffer.tinnews.R;
 import com.laioffer.tinnews.profile.TinProfileFragment;
