@@ -2,8 +2,8 @@ package com.laioffer.tinnews.common;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.LayoutRes;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.LayoutRes;
+import androidx.appcompat.app.AppCompatActivity;
 
 public abstract class TinBasicActivity extends AppCompatActivity implements TinFragmentManager {
     protected static final String BUNDLE = "bundle";
