@@ -1,8 +1,8 @@
 package com.laioffer.tinnews.common;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 public class TinFragmentPagerAdapter extends FragmentPagerAdapter {
     private Fragment[] fragments = new Fragment[FRAGMENT_NUMBER];
