@@ -1,7 +1,7 @@
 package com.laioffer.tinnews.mvp;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+import androidx.annotation.Nullable;
 
 import com.laioffer.tinnews.common.TinBasicFragment;
 
