@@ -2,20 +2,18 @@ package com.laioffer.tinnews.save;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
 
 import com.laioffer.tinnews.R;
-import com.laioffer.tinnews.common.TinBasicFragment;
 import com.laioffer.tinnews.common.ViewModelAdapter;
 import com.laioffer.tinnews.mvp.MvpFragment;
 import com.laioffer.tinnews.retrofit.response.News;
-import com.laioffer.tinnews.save.detail.SavedNewsDetailedFragment;
 
 import java.util.LinkedList;
 import java.util.List;
