@@ -2,7 +2,7 @@ package com.laioffer.tinnews.save.detail;
 
 import android.graphics.Paint;
 import android.os.Bundle;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.TextView;
 
