@@ -1,7 +1,7 @@
 package com.laioffer.tinnews.database;
 
-import android.arch.persistence.room.Database;
-import android.arch.persistence.room.RoomDatabase;
+import androidx.room.Database;
+import androidx.room.RoomDatabase;
 
 import com.laioffer.tinnews.retrofit.response.News;
 
